sandbox/docker/server/requirements.txt
grpcio-tools==1.74.0
grpcio==1.74.0
protobuf==6.31.1